flask_python_bypass_captcha/requirements.txt
flask>=2.3.1
Pillow==10.0.0